+++ get-libmame.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 ##
-## get-libmame.sh [target] [source]
+## get-libmame.sh [target] [clean | source]
 ##
 ## get version of libmame from MAME project, or the web
 ##
@@ -14,8 +14,8 @@
 ## source is path to MAME project or blank for default (../MAME)
 ##
 
-if [ "$1" == "" ]; then
-    $0 ios
+if [ "$1" == "" ] || [ "$1" == "clean" ]; then
+    $0 ios $1
     exit
 fi
 
@@ -26,7 +26,7 @@ if [ "$1" == "all" ]; then
     exit
 fi
 
-if [ "$1" != "ios" ] && [ "$1" != "tvos" ] && [ "$1" != "mac" ]; then
+if [ "$1" != "ios" ] && [ "$1" != "tvos" ] && [ "$1" != "mac" ] && [ "$1" != "ios-simulator" ] && [ "$1" != "tvos-simulator" ]; then
     echo "USAGE: $0 [ios | tvos | mac | all] [path to MAME]"
     exit
 fi
@@ -34,12 +34,26 @@ fi
 LIBMAME="libmame-$1.a"
 LIBMAME_URL="https://github.com/ToddLa/mame/releases/latest/download/$LIBMAME.gz"
 
-if [ -d "$2" ]; then
+## only do a clean and get out
+if [ "$2" == "clean" ]; then
+    echo CLEAN $LIBMAME
+    rm -f $LIBMAME
+    exit
+fi
+
+## detect if we are run from Xcode and need to do a CLEAN first
+if [ "TARGET_BUILD_DIR" != "" ] && [ $(ls -1 "$TARGET_BUILD_DIR" | wc -l) -lt 2 ]; then
+    rm -f $LIBMAME
+fi
+
+## copy from local custom build
+if [ -f "$2/$LIBMAME" ]; then
     echo COPY "$2/$LIBMAME"
     cp "$2/$LIBMAME" .
     exit
 fi
 
+## copy from local custom build
 if [ -f "../MAME/$LIBMAME" ]; then
     echo COPY "../MAME/$LIBMAME"
     cp "../MAME/$LIBMAME" .
@@ -47,7 +61,8 @@ if [ -f "../MAME/$LIBMAME" ]; then
 fi
 
 ## download from GitHub if no local version
-echo DOWNLOAD $LIBMAME
-curl -L $LIBMAME_URL | gunzip > $LIBMAME
-
+if [ ! -f "$LIBMAME" ]; then
+    echo DOWNLOAD $LIBMAME
+    curl -L $LIBMAME_URL | gunzip > $LIBMAME || (rm -r $LIBMAME; echo "DOWNLOAD $LIBMAME ** FAILED")
+fi