+++ make-ios.sh
@@ -1,11 +1,13 @@
 #!/bin/sh
-if [ "$1" != "noclean" ]; then
+
+if [ "$1" == "simulator" ]; then
+    export iOSSIMULATOR=1
+    shift
+fi
+
+## detect if we are run from Xcode and need to do a CLEAN first
+if [ "TARGET_BUILD_DIR" != "" ] && [ $(ls -1 "$TARGET_BUILD_DIR" | wc -l) -lt 2 ]; then
     make clean
 fi
 
-make -j`sysctl -n hw.logicalcpu` CDBG=-w || exit -1
-
-if [ "$1" != "noclean" ]; then
-    open xcode/MAME4iOS/MAME4iOS.xcodeproj/
-fi
-
+make -j`sysctl -n hw.logicalcpu` CDBG=-w $@ || exit -1