+++ WHATSNEW.md
@@ -5,6 +5,7 @@
 * Handle romsets names that end with `dash` number. (like `packman - 3.zip`)
 * Dont conflict with system keyboard shortcut for Prefrences `CMD+,`
 * Handle system keyboard shortcut for Escape `CMD+.`
+* Build LIBMAME via Xcode, no more need to run scripts from `Terminal.app`
 
 # Version 2021.7
 **NOTE** `MAME4iOS` 2021.7 comes in two versions, one that uses `MAME 139u1`, and one that uses the latest `MAME` (currently 234).  Make sure you download the version that is compatible with your ROMs, You can get the current version from the `Settings` page. If you want to use software based (ie cartridge, etc) romsets, please read the **Software Lists** section in the HELP or README.